Map generation, implementation of super sail power up
Implementation of the final map without colors and textures. Super sail power up implemented, i've added to the script Crest Boat Probes two methods: one that return the actual speed of the ship and one to set the new speed. The script for the power up is Power up speed. There is also a second power up, that does NOT work: barrels power up. Don't consider it.