Power up 3 (cannon faster reload), cannon fire effect (without sound))
Implementation of power up 3, the reload of the cannons goes from 8 to 3, after 30 seconds goes back to 8. Cannon explosion effect implementation. Anchor works but in the console there is an error(that does not affect nothing), an anchor spawn on the ship after 10 space press. To be fixed: the pirates when the ship is turned, move due to a centrifugal force.